cheri: initrd: Use a capability for initrd_start
Use uintptr_t for initrd_start and initrd_end.

+++ init/do_mounts_initrd.c
@@ -6,7 +6,7 @@
 
 #include "do_mounts.h"
 
-unsigned long initrd_start, initrd_end;
+uintptr_t initrd_start, initrd_end;
 int initrd_below_start_ok;
 static int __initdata mount_initrd = 1;
